Log runtime information for Eastwood process
In this commit, we are having the sonar-scanner process log
runtime information for the eastwood process eg. versions of Eastwood,
Clojure and JVM.

This info is included in the Eastwood output as the first line. We are
just extracting it from stdout and logging it. The log level is INFO
as this info can be handy for verifying that eastwood is being run on
the correct environment.