5. Add Cloverage support (#18)
* Remove testing extensions by number

Extensios adding is easier if there is not a test which requires certain number of extensions to be present. Easier is to add tests one by one

* Load property definition in first extension

Current test assumes propertydefinition to be in certain position. If position is changed to first this test cannot break when adding new extensions

* Move rules checking to sensors test

A sensor test should not assume certain numbers of rules to be present to avoid collisions with other sensors. Instead test should assert that rules that are used by sensor are in the rules not that rules are expected to contain only sensors rules.

* Add instructions to run SonarQube locally in docker

The script is added which builds the plugin and starts SonarQube with plugin installed in local environment. This is useful fo locally testing the components

* Change readme to have sensors in their own chapters

* Add Cloverage support

* Add Cloverage support

Cloverage is run via Leiningen. The coverage information is read and parsed from codecov.json file. Result is few percents differet than what Cloverage shows. This is most likely caused by different way to calculate coverage. Also SonarQube doesnt support partial hits.

* Refactor Cloverage sensor to use AbstractSensor

Also implementation has been changed to match other sensors

* Add pattern find to Cloverage report parsing

Cloverage codecov.json does not contain full relative path to file - the source root is omitted. File must therefore be found
by pattern matching against the sources

Code quality improvements

Update README.md

Co-Authored-By: hjhamala <heikki.hamalainen@iki.fi>

* Update README.md

Co-Authored-By: hjhamala <heikki.hamalainen@iki.fi>