4. Kibit support (#17)
* Remove testing extensions by number

Extensios adding is easier if there is not a test which requires certain number of extensions to be present. Easier is to add tests one by one

* Load property definition in first extension

Current test assumes propertydefinition to be in certain position. If position is changed to first this test cannot break when adding new extensions

* Move rules checking to sensors test

A sensor test should not assume certain numbers of rules to be present to avoid collisions with other sensors. Instead test should assert that rules that are used by sensor are in the rules not that rules are expected to contain only sensors rules.

* Add instructions to run SonarQube locally in docker

The script is added which builds the plugin and starts SonarQube with plugin installed in local environment. This is useful fo locally testing the components

* Change readme to have sensors in their own chapters

* Add support for Kibit

Kibit is run via Leiningen. Suggestions are mapped against correct lines. Kibit suggestion is by default minor code smell

* Add Kibit to readme

* Add Kibit support