Allow SonarClojure to work with SonarQube 7.1+
This commit fixes an issue related to the total lines of code that was
generating empty reports on SonarQube 7.1+. At the moment,
the LOC number is not accurate, but it is close enough to let the
plugin work fine.
In addition, we updated the lein plugin dependencies to their
latest versions.